+++ src/candlestick_chart/chart_data.py
@@ -36,6 +36,8 @@ class ChartData:
 
     def compute_visible_candles(self) -> None:
         nb_visible_candles = self.width - constants.WIDTH
+        if not constants.Y_AXIS_ON_THE_RIGHT:
+            nb_visible_candles -= constants.MARGIN_RIGHT
         self.visible_candle_set.set_candles(
             self.main_candle_set.candles[-nb_visible_candles:][:]
         )